requirements.txt
flet
pypdf
google-generativeai
python-dotenv
